% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{mg_send_email}
\alias{mg_send_email}
\title{Send Email with mailgun API}
\usage{
mg_send_email(from, to, cc = NULL, bcc = NULL, subject = NULL,
  text = NULL, html = NULL, attachment = NULL, inline = NULL,
  tag = NULL, campaign = NULL, dkim = NULL, deliverytime = NULL,
  testmode = NULL, tracking = NULL, tracking_clicks = NULL,
  tracking_opens = NULL, required_tls = NULL, skip_verification = NULL,
  x_my_header = NULL, my_var = NULL,
  mailgun_key = Sys.getenv("mailgun_key"),
  mailgun_domain = Sys.getenv("mailgun_domain"))
}
\description{
Send Email with mailgun API
}
